fix(Config): initialize feature flags in the configuration page
