+++ src/Components/AppConfig/AppConfig.tsx
@@ -19,6 +19,7 @@ import { Alert, Button, Checkbox, Field, FieldSet, Input, useStyles2 } from '@gr
 
 import { logger } from '../../services/logger';
 import { getDefaultDatasourceFromDatasourceSrv, getLastUsedDataSourceFromStorage } from '../../services/store';
+import { FeatureFlagContext } from 'Components/FeatureFlagContext';
 import { isValidTimeRange } from 'services/utils';
 
 export type JsonData = {
@@ -139,195 +140,197 @@ const AppConfig = ({ plugin }: Props) => {
     !state.defaultTimeRangeEnabled || (defaultTimeRangeValidation !== null && defaultTimeRangeValidation.valid);
 
   return (
-    <div data-testid={testIds.appConfig.container}>
-      <FieldSet label={t('components.app-config.label-settings', 'Settings')}>
-        <Field
-          description={
-            <span>
-              <Trans i18nKey="components.app-config.default-data-source-description">
-                The default data source to be used for new Logs Drilldown users. Each user can override their default by
-                setting another data source in Logs Drilldown.
-              </Trans>
-            </span>
-          }
-          label={t('components.app-config.label-default-data-source', 'Default data source')}
-        >
-          <DataSourcePicker
-            width={60}
-            filter={(ds) => ds.type === 'loki'}
-            current={state.dataSource}
-            onChange={onChangeDatasource}
-          />
-        </Field>
-
-        <Field
-          className={styles.marginTop}
-          description={
-            <span>
-              <Trans i18nKey="components.app-config.default-time-range-description">
-                When enabled, this time range is used when users open Logs Drilldown for the first time, and without a
-                time range in the URL. When disabled, the app uses its built-in default (last 15 minutes).
-              </Trans>
-            </span>
-          }
-          label={t('components.app-config.label-default-time-range', 'Default time range')}
-        >
-          <Checkbox
-            id="default-time-range-enabled"
-            data-testid={testIds.appConfig.defaultTimeRangeEnabled}
-            label={t(
-              'components.app-config.default-time-range-enabled-label-use-custom-default-time-range',
-              'Use custom default time range'
-            )}
-            value={state.defaultTimeRangeEnabled}
-            onChange={onChangeDefaultTimeRangeEnabled}
-          />
-        </Field>
-
-        {state.defaultTimeRangeEnabled && (
-          <div className={styles.defaultTimeRangeInputs}>
-            <Field
-              invalid={defaultTimeRangeValidation !== null && !defaultTimeRangeValidation.valid}
-              error={
-                defaultTimeRangeValidation !== null && !defaultTimeRangeValidation.valid
-                  ? defaultTimeRangeValidation.error
-                  : undefined
-              }
-              description={t(
-                'components.app-config.description-start-range',
-                'Start of the range (e.g. now-15m, now-1h, now-24h)'
-              )}
-              label={t('components.app-config.label-from', 'From')}
-            >
-              <Input
-                width={40}
-                id="default-time-range-from"
-                data-testid={testIds.appConfig.defaultTimeRangeFrom}
-                value={state.defaultTimeRangeFrom}
-                placeholder={t('components.app-config.default-time-range-from-placeholder-now-15m', 'now-15m')}
-                onChange={onChangeDefaultTimeRangeFrom}
-              />
-            </Field>
-            <Field
-              className={styles.marginTop}
-              invalid={defaultTimeRangeValidation !== null && !defaultTimeRangeValidation.valid}
-              error={
-                defaultTimeRangeValidation !== null && !defaultTimeRangeValidation.valid
-                  ? defaultTimeRangeValidation.error
-                  : undefined
-              }
-              description={t(
-                'components.app-config.description-end-of-the-range-eg-now',
-                'End of the range (e.g. now)'
-              )}
-              label={t('components.app-config.label-to', 'To')}
-            >
-              <Input
-                width={40}
-                id="default-time-range-to"
-                data-testid={testIds.appConfig.defaultTimeRangeTo}
-                value={state.defaultTimeRangeTo}
-                placeholder={t('components.app-config.default-time-range-to-placeholder-now', 'now')}
-                onChange={onChangeDefaultTimeRangeTo}
-              />
-            </Field>
-          </div>
-        )}
-
-        <Field
-          invalid={!isValid(state.interval)}
-          error={t(
-            'components.app-config.interval-invalid-error',
-            'Interval is invalid. Please enter an interval longer than "60m". For example: 3d, 1w, 1m'
-          )}
-          description={
-            <span>
-              <Trans i18nKey="components.app-config.max-interval-description">
-                The maximum interval that can be selected in the time picker within the Grafana Logs Drilldown app. If
-                empty, users can select any time range interval in Grafana Logs Drilldown. <br />
-                Example values: 7d, 24h, 2w
-              </Trans>
-            </span>
-          }
-          label={t('components.app-config.label-maximum-time-picker-interval', 'Maximum time picker interval')}
-          className={styles.marginTop}
-        >
-          <Input
-            width={60}
-            id="interval"
-            data-testid={testIds.appConfig.interval}
-            label={t('components.app-config.label-max-interval', 'Max interval')}
-            value={state?.interval}
-            placeholder={t('components.app-config.interval-placeholder', '7d')}
-            onChange={onChangeInterval}
-          />
-        </Field>
-
-        <Field
-          className={styles.marginTop}
-          description={
-            <span>
-              <Trans i18nKey="components.app-config.disable-patterns-description">
-                Disables Logs Drilldown&apos;s usage of the{' '}
-                <a
-                  className="external-link"
-                  href="https://grafana.com/docs/loki/latest/reference/loki-http-api/#patterns-detection"
-                  target="_blank"
-                  rel="noreferrer"
-                >
-                  Loki Patterns API
-                </a>{' '}
-                endpoint, and removes the Patterns tab.
-              </Trans>
-            </span>
-          }
-          label={t('components.app-config.label-disable-loki-patterns', 'Disable Loki patterns')}
-        >
-          <Checkbox
-            id="disable-patterns"
-            data-testid={testIds.appConfig.interval}
-            label={t('components.app-config.label-disable-patterns', 'Disable patterns')}
-            value={state?.patternsDisabled}
-            placeholder={t('components.app-config.patterns-placeholder', '7d')}
-            onChange={onChangePatternsDisabled}
-          />
-        </Field>
-
-        <div className={styles.marginTop}>
-          <Button
-            type="submit"
-            data-testid={testIds.appConfig.submit}
-            onClick={() =>
-              updatePluginAndReload(plugin.meta.id, {
-                enabled,
-                jsonData: {
-                  dataSource: state.dataSource,
-                  interval: state.interval,
-                  patternsDisabled: state.patternsDisabled,
-                  defaultTimeRange:
-                    state.defaultTimeRangeEnabled &&
-                    state.defaultTimeRangeFrom.trim() &&
-                    state.defaultTimeRangeTo.trim()
-                      ? { from: state.defaultTimeRangeFrom.trim(), to: state.defaultTimeRangeTo.trim() }
-                      : undefined,
-                },
-                pinned,
-              })
+    <FeatureFlagContext>
+      <div data-testid={testIds.appConfig.container}>
+        <FieldSet label={t('components.app-config.label-settings', 'Settings')}>
+          <Field
+            description={
+              <span>
+                <Trans i18nKey="components.app-config.default-data-source-description">
+                  The default data source to be used for new Logs Drilldown users. Each user can override their default
+                  by setting another data source in Logs Drilldown.
+                </Trans>
+              </span>
             }
-            disabled={!isValid(state.interval) || !isDefaultTimeRangeValid}
+            label={t('components.app-config.label-default-data-source', 'Default data source')}
           >
-            <Trans i18nKey="components.app-config.save-settings">Save settings</Trans>
-          </Button>
-        </div>
-        <div className={styles.note}>
-          <Alert severity="info" title="">
-            <Trans i18nKey="components.app-config.active-users-reload-reflect-configuration-changes">
-              Active users must reload the app to reflect configuration changes.
-            </Trans>
-          </Alert>
-        </div>
-      </FieldSet>
-    </div>
+            <DataSourcePicker
+              width={60}
+              filter={(ds) => ds.type === 'loki'}
+              current={state.dataSource}
+              onChange={onChangeDatasource}
+            />
+          </Field>
+
+          <Field
+            className={styles.marginTop}
+            description={
+              <span>
+                <Trans i18nKey="components.app-config.default-time-range-description">
+                  When enabled, this time range is used when users open Logs Drilldown for the first time, and without a
+                  time range in the URL. When disabled, the app uses its built-in default (last 15 minutes).
+                </Trans>
+              </span>
+            }
+            label={t('components.app-config.label-default-time-range', 'Default time range')}
+          >
+            <Checkbox
+              id="default-time-range-enabled"
+              data-testid={testIds.appConfig.defaultTimeRangeEnabled}
+              label={t(
+                'components.app-config.default-time-range-enabled-label-use-custom-default-time-range',
+                'Use custom default time range'
+              )}
+              value={state.defaultTimeRangeEnabled}
+              onChange={onChangeDefaultTimeRangeEnabled}
+            />
+          </Field>
+
+          {state.defaultTimeRangeEnabled && (
+            <div className={styles.defaultTimeRangeInputs}>
+              <Field
+                invalid={defaultTimeRangeValidation !== null && !defaultTimeRangeValidation.valid}
+                error={
+                  defaultTimeRangeValidation !== null && !defaultTimeRangeValidation.valid
+                    ? defaultTimeRangeValidation.error
+                    : undefined
+                }
+                description={t(
+                  'components.app-config.description-start-range',
+                  'Start of the range (e.g. now-15m, now-1h, now-24h)'
+                )}
+                label={t('components.app-config.label-from', 'From')}
+              >
+                <Input
+                  width={40}
+                  id="default-time-range-from"
+                  data-testid={testIds.appConfig.defaultTimeRangeFrom}
+                  value={state.defaultTimeRangeFrom}
+                  placeholder={t('components.app-config.default-time-range-from-placeholder-now-15m', 'now-15m')}
+                  onChange={onChangeDefaultTimeRangeFrom}
+                />
+              </Field>
+              <Field
+                className={styles.marginTop}
+                invalid={defaultTimeRangeValidation !== null && !defaultTimeRangeValidation.valid}
+                error={
+                  defaultTimeRangeValidation !== null && !defaultTimeRangeValidation.valid
+                    ? defaultTimeRangeValidation.error
+                    : undefined
+                }
+                description={t(
+                  'components.app-config.description-end-of-the-range-eg-now',
+                  'End of the range (e.g. now)'
+                )}
+                label={t('components.app-config.label-to', 'To')}
+              >
+                <Input
+                  width={40}
+                  id="default-time-range-to"
+                  data-testid={testIds.appConfig.defaultTimeRangeTo}
+                  value={state.defaultTimeRangeTo}
+                  placeholder={t('components.app-config.default-time-range-to-placeholder-now', 'now')}
+                  onChange={onChangeDefaultTimeRangeTo}
+                />
+              </Field>
+            </div>
+          )}
+
+          <Field
+            invalid={!isValid(state.interval)}
+            error={t(
+              'components.app-config.interval-invalid-error',
+              'Interval is invalid. Please enter an interval longer than "60m". For example: 3d, 1w, 1m'
+            )}
+            description={
+              <span>
+                <Trans i18nKey="components.app-config.max-interval-description">
+                  The maximum interval that can be selected in the time picker within the Grafana Logs Drilldown app. If
+                  empty, users can select any time range interval in Grafana Logs Drilldown. <br />
+                  Example values: 7d, 24h, 2w
+                </Trans>
+              </span>
+            }
+            label={t('components.app-config.label-maximum-time-picker-interval', 'Maximum time picker interval')}
+            className={styles.marginTop}
+          >
+            <Input
+              width={60}
+              id="interval"
+              data-testid={testIds.appConfig.interval}
+              label={t('components.app-config.label-max-interval', 'Max interval')}
+              value={state?.interval}
+              placeholder={t('components.app-config.interval-placeholder', '7d')}
+              onChange={onChangeInterval}
+            />
+          </Field>
+
+          <Field
+            className={styles.marginTop}
+            description={
+              <span>
+                <Trans i18nKey="components.app-config.disable-patterns-description">
+                  Disables Logs Drilldown&apos;s usage of the{' '}
+                  <a
+                    className="external-link"
+                    href="https://grafana.com/docs/loki/latest/reference/loki-http-api/#patterns-detection"
+                    target="_blank"
+                    rel="noreferrer"
+                  >
+                    Loki Patterns API
+                  </a>{' '}
+                  endpoint, and removes the Patterns tab.
+                </Trans>
+              </span>
+            }
+            label={t('components.app-config.label-disable-loki-patterns', 'Disable Loki patterns')}
+          >
+            <Checkbox
+              id="disable-patterns"
+              data-testid={testIds.appConfig.interval}
+              label={t('components.app-config.label-disable-patterns', 'Disable patterns')}
+              value={state?.patternsDisabled}
+              placeholder={t('components.app-config.patterns-placeholder', '7d')}
+              onChange={onChangePatternsDisabled}
+            />
+          </Field>
+
+          <div className={styles.marginTop}>
+            <Button
+              type="submit"
+              data-testid={testIds.appConfig.submit}
+              onClick={() =>
+                updatePluginAndReload(plugin.meta.id, {
+                  enabled,
+                  jsonData: {
+                    dataSource: state.dataSource,
+                    interval: state.interval,
+                    patternsDisabled: state.patternsDisabled,
+                    defaultTimeRange:
+                      state.defaultTimeRangeEnabled &&
+                      state.defaultTimeRangeFrom.trim() &&
+                      state.defaultTimeRangeTo.trim()
+                        ? { from: state.defaultTimeRangeFrom.trim(), to: state.defaultTimeRangeTo.trim() }
+                        : undefined,
+                  },
+                  pinned,
+                })
+              }
+              disabled={!isValid(state.interval) || !isDefaultTimeRangeValid}
+            >
+              <Trans i18nKey="components.app-config.save-settings">Save settings</Trans>
+            </Button>
+          </div>
+          <div className={styles.note}>
+            <Alert severity="info" title="">
+              <Trans i18nKey="components.app-config.active-users-reload-reflect-configuration-changes">
+                Active users must reload the app to reflect configuration changes.
+              </Trans>
+            </Alert>
+          </div>
+        </FieldSet>
+      </div>
+    </FeatureFlagContext>
   );
 };
